requirements.txt
fastapi~=0.115.13
uvicorn~=0.34.3
sqlalchemy~=2.0.41
psycopg2-binary
python-dotenv~=1.1.0
email-validator
pytest~=8.4.1
httpx

pydantic~=2.11.7